Delete extra space, rename files JC-551
