+++ apps/angular/src/app/features/auth/login-page/login-page.component.ts
@@ -8,7 +8,7 @@ import { BehaviorSubject, finalize, first, tap } from 'rxjs';
 import { untilDestroyed } from '@js-camp/angular/core/rxjs/until-destroyed';
 import { AppError, AppErrors } from '@js-camp/core/models/app-error';
 import { AppValidators } from '@js-camp/angular/core/utils/validators';
-import { catchHttpErrorResponse } from '@js-camp/angular/core/rxjs/catchHttpErrorResponse';
+import { catchHttpErrorResponse } from '@js-camp/angular/core/rxjs/catch-http-error-response';
 
 const defaultFormValues: Login = {
 	email: '',
